+++ FoodDlvAPI/Interfaces/ICartRepository.cs
@@ -10,21 +10,21 @@ namespace FoodDlvAPI.Interfaces
         /// </summary>
         /// <param name="memberId"></param>
         /// <returns></returns>
-        bool IsExists(long memberId, int storeId);
+        bool IsExists(int memberId, int storeId);
 
         /// <summary>
         /// 建立該會員對應商店專屬的空的購物車
         /// </summary>
         /// <param name="memberId"></param>
         /// <returns></returns>
-        CartDTO CreateNewCart(long memberId, int storeId);
+        CartDTO CreateNewCart(int memberId, int storeId);
 
         /// <summary>
         /// 讀取目前購物車的必要內容
         /// </summary>
         /// <param name="memberId"></param>
         /// <returns></returns>
-        CartDTO Load(long memberId, int storeId);
+        CartDTO Load(int memberId, int storeId);
 
         /// <summary>
         /// 添加購物車明細, 包含各項驗證功能
@@ -37,7 +37,7 @@ namespace FoodDlvAPI.Interfaces
         /// 清空購物車, 並刪除Database中的購物車主檔與明細檔
         /// </summary>
         /// <param name="memberId"></param>
-        void EmptyCart(long memberId, int storeId);
+        void EmptyCart(int memberId, int storeId);
 
         /// <summary>
         /// 監聽"加入購物車", 觸發商品客製化識別號碼的選擇器
