+++ FoodDlvAPI/Repositories/CartRepository.cs
@@ -20,7 +20,7 @@ namespace FoodDlvAPI.Repositories
             _context = context;
         }
 
-        public bool IsExists(long memberId, int storeId)
+        public bool IsExists(int memberId, int storeId)
         {
             var memberCheck = _context.Members.Any(m => m.Id == memberId);
             var storeCheck = _context.Stores.Any(s => s.Id == storeId);
@@ -36,7 +36,7 @@ namespace FoodDlvAPI.Repositories
             }
         }
 
-        public CartDTO Load(long memberId, int storeId)
+        public CartDTO Load(int memberId, int storeId)
         {
             var data = _context.Carts
                 .AsNoTracking()
@@ -47,7 +47,7 @@ namespace FoodDlvAPI.Repositories
             return data;
         }
 
-        public CartDTO CreateNewCart(long memberId, int storeId)
+        public CartDTO CreateNewCart(int memberId, int storeId)
         {
             var cart = new Cart
             {
@@ -60,7 +60,7 @@ namespace FoodDlvAPI.Repositories
             return Load(memberId, storeId);
         }
 
-        public void EmptyCart(long memberId, int storeId)
+        public void EmptyCart(int memberId, int storeId)
         {
             var cart = _context.Carts.SingleOrDefault(c => c.MemberId == memberId && c.StoreId == storeId);
             if (cart == null) return;
